Move SimpleLPPrefix to DEA and remove readmes
