--- lib/dea/Readme.md
@@ -1,3 +1,0 @@
-
-DEA solver
-==========
