+++ lib/dea/src/main/java/dk/alexandra/fresco/lib/statistics/DEAPrefixBuilderMaximize.java
@@ -7,7 +7,6 @@ import dk.alexandra.fresco.framework.builder.numeric.ProtocolBuilderNumeric;
 import dk.alexandra.fresco.framework.value.SInt;
 import dk.alexandra.fresco.lib.collections.Matrix;
 import dk.alexandra.fresco.lib.lp.LPTableau;
-import dk.alexandra.fresco.lib.lp.SimpleLPPrefix;
 
 import java.math.BigInteger;
 import java.util.ArrayList;
